Finished stringBuilder in c.
