+++ dataStuctures/c/stringBuilder/Makefile
@@ -1,0 +1,3 @@
+linkedList: stringBuilder.c main.c
+	gcc -g -o stringBuilder ../linkedList/linkedList.c stringBuilder.c main.c
+
