+++ dataStuctures/c/stringBuilder/valgrind.sh
@@ -1,4 +1,4 @@
 # Quickly test and run valgrind, checks for memory issues
-gcc -g ../linkedList/linkedList.c stringBuilder.c
+gcc -g ../linkedList/linkedList.c stringBuilder.c main.c
 valgrind ./a.out
 rm a.out
